Make state a global object and get referenes
Previously the state was passed in a functional way. This would not be
feasible if we use game buttons instead of console. So we must maintain
state as a global variable.

I added the 3 game buttons and got references in script to prepare to
assign playRound logic to each of them.